# api/openapi.yaml
openapi: 3.1.0
info:
  title: Speccy API
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
paths:
...
  /handoffs/{handoffId}/report:
    parameters:
      - $ref: "#/components/parameters/HandoffId"
    post:
      operationId: reportBuild
      summary: Report what the build learned about the doc (REQ-137). A blocked report opens a blocking thread.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildReport"
      responses:
        "200":
          description: The thread the report opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetail"
        default:
          $ref: "#/components/responses/Problem"
...
  /bundles:
    get:
      operationId: listBundles
      summary: List bundles, and the folders that look like bundles but are not valid.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: One page of bundles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleList"
        default:
          $ref: "#/components/responses/Problem"

    post:
      operationId: createBundle
      summary: Create a bundle from a profile's template (REQ-016).
      description: In local mode, Speccy writes the bundle to a new folder under the folder that Speccy serves.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBundleRequest"
      responses:
        "201":
          description: The new bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDoc"
        default:
          $ref: "#/components/responses/Problem"
...
  /bundles/{bundleId}:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    get:
      operationId: getBundle
      summary: Get one bundle with its spec docs.
      responses:
        "200":
          description: The bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteBundle
      summary: Delete a bundle whose text Speccy holds, with everything that hangs off it. Permanent.
      parameters:
        - name: slug
          in: query
          required: true
          description: The bundle's slug, typed to confirm.
          schema: { type: string }
      responses:
        "204":
          description: The bundle is gone.
        default:
          $ref: "#/components/responses/Problem"
...
  /reviews:
    post:
      operationId: reviewContent
      summary: Review bundle files that are not saved on the server (SDD §12.2 --server, REQ-111 review_content). The server keeps the files and the result for 90 days for the report (SDD §12.4); it changes no bundle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentReviewRequest"
      responses:
        "200":
          description: The findings and the verdict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentReview"
        default:
          $ref: "#/components/responses/Problem"
...
  /docs/{docId}/tour:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: getTour
      summary: The ordered points of the bundle's current review that need a human decision (SDD §13.3).
      responses:
        "200":
          description: The tour.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        default:
          $ref: "#/components/responses/Problem"
...
  /threads/{threadId}/messages:
    parameters:
      - $ref: "#/components/parameters/ThreadId"
    post:
      operationId: postMessage
      summary: Post a message. In a thread for the AI, the AI answers with sources (REQ-088).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostMessage"
      responses:
        "200":
          description: The thread.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetail"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
...
    ThreadId:
      name: threadId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BundleId:
      name: bundleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocId:
      name: docId
      in: path
      required: true
      description: The ID of one spec doc.
      schema:
        type: string
        format: uuid
    HandoffId:
      name: handoffId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  responses:
    Problem:
      description: An error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details with a stable code.
      required: [type, title, status, code]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: A stable machine-readable error code.
    SpecDoc:
      type: object
      description: One spec doc of a bundle, with its own profile, versions, review runs and verdict.
      required: [id, bundle_id, slug, title, profile_key, source_kind, path, current_version, updated_at]
      properties:
        id:
          type: string
          format: uuid
        bundle_id:
          type: string
          format: uuid
          description: The bundle that holds the spec doc.
        slug:
          type: string
        title:
          type: string
        profile_key:
          type: string
          description: The profile of the spec doc.
        source_kind:
          type: string
          enum: [local, db, github]
        github:
          $ref: "#/components/schemas/BundleGithub"
        path:
          type: string
          description: The path of the spec doc in its bundle.
        current_version:
          $ref: "#/components/schemas/Version"
        updated_at:
          type: string
          format: date-time
        verdict:
          $ref: "#/components/schemas/BundleVerdict"
        run_error:
          type: string
          description: Why the latest run on the current version failed, when it failed.
        adopt:
          $ref: "#/components/schemas/Adopt"
        status:
          $ref: "#/components/schemas/ReviewStatus"
        next_action:
          $ref: "#/components/schemas/NextAction"
    Bundle:
      type: object
      description: A folder that holds one or more spec docs and their assets.
      required: [id, slug, title, source_kind, state, docs, updated_at]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          description: The bundle folder relative to the served folder or the repo root.
        title:
          type: string
          description: The folder name.
        source_kind:
          type: string
          enum: [local, db, github]
        visibility:
          $ref: "#/components/schemas/Visibility"
        state:
          $ref: "#/components/schemas/BundleState"
        docs:
          type: array
          description: The spec docs of the bundle, by path.
          items:
            $ref: "#/components/schemas/SpecDoc"
        updated_at:
          type: string
          format: date-time
...
    CreateBundleRequest:
      type: object
      required: [profile, name]
      properties:
        profile:
          type: string
          description: The profile key.
        name:
          type: string
          description: The bundle folder name.
        title:
          type: string
...
    Tour:
      type: object
      required: [points]
      properties:
        run_id:
          type: string
          format: uuid
          description: The run whose findings the tour uses. Absent before the first review.
        points:
          type: array
          items:
            $ref: "#/components/schemas/TourPoint"
...
    ContentReviewRequest:
      type: object
      required: [files]
      properties:
        slug:
          type: string
          description: The bundle's slug. Links from other bundles and to this one use it.
        main_doc:
          type: string
          description: The main doc of a single-file bundle (REQ-001 form b). Absent means the one file with a type in its frontmatter.
        profile:
          type: string
          description: The profile for a main doc with no type in its frontmatter (REQ-130).
        stages:
          type: array
          items:
            type: string
            enum: [rubric, grounding, divergence, coherence]
        files:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/ContentFile"
    ContentReview:
      type: object
      required: [id, report_path, profile_key, profile_version, main_doc, verdict, findings, notes]
      properties:
        id:
          type: string
          format: uuid
        report_path:
          type: string
          description: The app page of the report, relative to the server, such as /reviews/{id}.
        size:
          type: string
          description: The doc size the review used, from the frontmatter or inferred (REQ-134).
        profile_key:
          type: string
        profile_version:
          type: integer
          format: int64
        main_doc:
          type: string
        verdict:
          $ref: "#/components/schemas/ContentVerdict"
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        notes:
          type: array
          items:
            type: string
        tokens_in:
          type: integer
          format: int64
        tokens_out:
          type: integer
          format: int64
        cost_estimate:
          type: number
        cache_hits:
          type: integer
...
    Thread:
      type: object
      required: [id, anchor_kind, anchor, addressed_to, title, blocking, status, created_by, created_at, last_message_at, message_count]
      properties:
        id:
          type: string
          format: uuid
        doc_id:
          type: string
          format: uuid
        profile_key:
          type: string
        anchor_kind:
          type: string
          enum: [text, section, finding, check]
        anchor:
          type: object
          additionalProperties: true
          description: For text, an Anchor. For a section, {heading_path}. For a finding, {finding_id, check_slug}. For a check, {check_slug}.
        addressed_to:
          type: string
          enum: [humans, ai]
        title:
          type: string
        handoff_id:
          type: string
          format: uuid
          description: The handoff a builder opened this thread from (REQ-137).
        handoff_version:
          type: integer
          format: int64
          description: The bundle version the builder took. 0 when no handoff opened it.
        blocking:
          type: boolean
        status:
          type: string
          enum: [open, resolved]
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        last_message_at:
          type: string
          format: date-time
        message_count:
          type: integer
    ThreadDetail:
      allOf:
        - $ref: "#/components/schemas/Thread"
        - type: object
          required: [messages, answering]
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/ThreadMessage"
            answering:
              type: boolean
              description: The AI is writing an answer.
    ThreadMessage:
      type: object
      required: [id, seq, author_kind, author_name, body, sources, decision, created_at]
      properties:
        id:
          type: string
          format: uuid
        seq:
          type: integer
        author_kind:
          type: string
          enum: [user, guest, ai]
        author_id:
          type: string
        author_name:
          type: string
        body:
          type: string
        sources:
          type: array
          items:
            type: string
        decision:
          type: string
          enum: ["", decision, reversal]
        created_at:
          type: string
          format: date-time
...
    PostMessage:
      type: object
      required: [body]
      properties:
        body:
          type: string
...
    BuildReport:
      type: object
      description: What a coding agent learned about the doc while it built from it (REQ-137).
      required: [kind, text]
      properties:
        kind:
          type: string
          enum: [blocked, note]
          description: blocked means the agent cannot build the section without an answer. note means it built something, and the doc was unclear.
        section:
          type: array
          description: The heading path of the section the report is about.
          items:
            type: string
        trace_id:
          type: string
          description: A trace ID the report is about, such as REQ-012. Speccy anchors the thread to where the doc defines it.
        text:
          type: string
          description: What the agent needs, in its own words.
...
    ReviewStatus:
      type: string
      enum: [draft, in_review, approved, superseded]
...
    BundleList:
      type: object
      required: [items, problems]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Bundle"
        next_cursor:
          type: string
        problems:
          type: array
          description: Folders that are not valid bundles, for example with two main docs (REQ-001).
          items:
            $ref: "#/components/schemas/BundleProblem"
...